+++ apps/formio-server/src/modules/modules.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = function(app, config) {
-  return {
-    websocket: require('./websocket')(app, config)
-  };
-};